requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.1
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.10
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==19.7.1
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.3
pipenv==2018.11.26
pipfile==0.0.2
psycopg2==2.7.5
psycopg2-binary==2.8.5
pycparser==2.20
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.6
pytz==2020.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.16
virtualenv-clone==0.5.4
whitenoise==3.3.1
wrapt==1.12.1
zipp==3.1.0
